% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASETS.R
\docType{package}
\name{AOC2019}
\alias{AOC2019}
\alias{Advent-Of-Code-2019}
\alias{AOC2019-package}
\title{Advent of Code Challenge 2019 Code Store}
\description{
This package contains functions, test, context and datasets for the AOC challenge 2019
}
