% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day07.R
\name{da07_add_value_to_list_elements}
\alias{da07_add_value_to_list_elements}
\title{add value into every element of list}
\usage{
da07_add_value_to_list_elements(list_of_values, value)
}
\arguments{
\item{list_of_values}{input list that needs to be extended with value}

\item{value}{value to all into every element of list}
}
\description{
This is supplementary function for permutations
}
