% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day01.R
\name{day01_fuelrec}
\alias{day01_fuelrec}
\title{fuel adjustment on fuel itself}
\usage{
day01_fuelrec(fuel_mass)
}
\arguments{
\item{fuel_mass}{mass of fuel that needs additional fuel}
}
\description{
fuel adjustment on fuel itself
}
