% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{day02_modify_input}
\alias{day02_modify_input}
\title{puzzle requires input modification}
\usage{
day02_modify_input(input, v1, v2)
}
\arguments{
\item{input}{input vector of integers}
}
\description{
puzzle requires input modification
}
