% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{day02_noun_verb}
\alias{day02_noun_verb}
\title{find combination of word and noun}
\usage{
day02_noun_verb(input, expected_result)
}
\arguments{
\item{input}{input}

\item{expected_result}{expected result}
}
\description{
find combination of word and noun
}
