% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\docType{data}
\name{day03_centralpoint}
\alias{day03_centralpoint}
\title{Central point - starting position}
\format{An object of class \code{numeric} of length 2.}
\usage{
day03_centralpoint
}
\description{
Central point - starting position
}
\keyword{datasets}
