% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{day03_deltapath}
\alias{day03_deltapath}
\title{build delta to the path according}
\usage{
day03_deltapath(current_pos, direction, steps)
}
\arguments{
\item{current_pos}{current position (to start from)}

\item{direction}{direction to move in}

\item{steps}{number of steps to make}
}
\description{
build delta to the path according
}
