% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{day03_intersection}
\alias{day03_intersection}
\title{Find all intersections of 2 pathes}
\usage{
day03_intersection(path1, path2)
}
\arguments{
\item{path1}{path}

\item{path2}{path}
}
\description{
Returns pairs of x and y coordinates in data frame
}
