% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{day03_mhtdist}
\alias{day03_mhtdist}
\title{Manhettan distance}
\usage{
day03_mhtdist(point, ref_point = c(0, 0))
}
\arguments{
\item{point}{pair numbers representing X and Y coordinate}

\item{ref_point}{pair numbers representing X and Y coordinate}
}
\description{
Manhettan distance
}
