% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{day03_part1}
\alias{day03_part1}
\title{Fund find closest (manhettan distance) intersection of 2 wires defined by instruction sets}
\usage{
day03_part1(instructions1, instructions2)
}
\arguments{
\item{instructions1}{instructions to build path 1}

\item{instructions2}{instructions to build path 2}
}
\description{
Fund find closest (manhettan distance) intersection of 2 wires defined by instruction sets
}
