% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{day03_part2}
\alias{day03_part2}
\title{Find shortest (wire length) intersection of 2 wires defined by instruction sequences}
\usage{
day03_part2(instructions1, instructions2)
}
\arguments{
\item{instructions1}{instructions to build path 1}

\item{instructions2}{instructions to build path 2}
}
\description{
Find shortest (wire length) intersection of 2 wires defined by instruction sequences
}
