% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{day03_path_direction}
\alias{day03_path_direction}
\title{Translate instruction into directoin vector (x, y)}
\usage{
day03_path_direction(instruction)
}
\arguments{
\item{instruction}{string that defines direction as a first character and
number of steps as rest}
}
\description{
R : (1, 0); U : (0, 1); L : (-1, 0); D : (0, -1)
}
