% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{day03_path_length}
\alias{day03_path_length}
\title{Translate instruction into number of steps}
\usage{
day03_path_length(instruction)
}
\arguments{
\item{instruction}{string that defines direction as a first character and
number of steps as rest}
}
\description{
Translate instruction into number of steps
}
