% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day04.R
\name{day04_docheck}
\alias{day04_docheck}
\title{Function that generalizes 2 types of checks using binary check operator and}
\usage{
day04_docheck(number, check_fun, fold_fun)
}
\arguments{
\item{number}{value to analyze}

\item{check_fun}{function to apply to check values}

\item{fold_fun}{function to apply to fold results of check}
}
\description{
function that folds outcomes to say whether check passes or not
}
