% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day04.R
\name{day04_filteradj}
\alias{day04_filteradj}
\title{Check whether number has at least to adjacent digits}
\usage{
day04_filteradj(number)
}
\arguments{
\item{number}{number to be checked}
}
\description{
where 00111 and 11322 are valid and 12321 is not. Function works with atomic
values only
}
