% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day04.R
\name{day04_filteradj2}
\alias{day04_filteradj2}
\title{Check whether adjacent digits are not part of bigger group}
\usage{
day04_filteradj2(number)
}
\arguments{
\item{number}{number to be checked}
}
\description{
where 00111 and 11322 are valid and 12321 is not. Function works with atomic
values only
}
