% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day04.R
\name{day04_filterasc}
\alias{day04_filterasc}
\title{Check whether digits of number all go in asscending order}
\usage{
day04_filterasc(number)
}
\arguments{
\item{number}{number to be checked}
}
\description{
Where 00111 and 12345 are valid and 12321 is not. Function works with atomic
values only
}
