% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day04.R
\name{day04_makesequence}
\alias{day04_makesequence}
\title{Generate sequence of numbers based on value of low and high bound}
\usage{
day04_makesequence(low, high)
}
\arguments{
\item{low}{low bound of range}

\item{high}{high bound of range}
}
\description{
Convert numbers into string format after validation
}
