% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05_changebase}
\alias{day05_changebase}
\title{change relative base}
\usage{
day05_changebase(pointer, base, array, in_buffer, out_buffer)
}
\arguments{
\item{pointer}{pointer}

\item{base}{relative base of pointer for parameter mode 2}

\item{array}{array}

\item{in_buffer}{input array}

\item{out_buffer}{output array}
}
\description{
change relative base
}
