% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05_getv}
\alias{day05_getv}
\title{Get value}
\usage{
day05_getv(mode, pointer, array)
}
\arguments{
\item{mode}{parameter mode (read from array or show exact value)}

\item{pointer}{pointer to element in array or exact value}

\item{array}{array of intcodes}
}
\description{
Get value
}
