% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05_instruct_par_mode}
\alias{day05_instruct_par_mode}
\title{instruction parameter modes}
\usage{
day05_instruct_par_mode(instruction)
}
\arguments{
\item{instruction}{instruction integer code}
}
\description{
instruction parameter modes
}
