% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05_intruct_length}
\alias{day05_intruct_length}
\title{Length of instruction - how many registers belong to instruction}
\usage{
day05_intruct_length(instruction)
}
\arguments{
\item{instruction}{instruction integer code}
}
\description{
Length of instruction - how many registers belong to instruction
}
