% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05_intruct_op}
\alias{day05_intruct_op}
\title{decode instruction to identify operation}
\usage{
day05_intruct_op(instruction)
}
\arguments{
\item{instruction}{instruction code: integer value or digital code}
}
\description{
decode instruction to identify operation
}
