% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05_oneifequal}
\alias{day05_oneifequal}
\title{move pointer if frist parameter is equal to second}
\usage{
day05_oneifequal(pointer, array, in_buffer, out_buffer)
}
\arguments{
\item{pointer}{pointer}

\item{array}{array}

\item{in_buffer}{input array}

\item{out_buffer}{output array}
}
\description{
move pointer if frist parameter is equal to second
}
