% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05_opmul}
\alias{day05_opmul}
\title{multiply two numbers}
\usage{
day05_opmul(pointer, base, array, in_buffer, out_buffer)
}
\arguments{
\item{pointer}{current position of instruction}

\item{base}{relative base of pointer for parameter mode 2}

\item{array}{incode array}

\item{in_buffer}{vector of values to be read}

\item{out_buffer}{vector of output}
}
\description{
multiply two numbers
}
