% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05_opoutput}
\alias{day05_opoutput}
\title{return result (print output)}
\usage{
day05_opoutput(pointer, array, in_buffer, out_buffer)
}
\arguments{
\item{pointer}{current position of instruction}

\item{array}{incode array}

\item{in_buffer}{vector of values to be read}

\item{out_buffer}{vector of output}
}
\description{
return result (print output)
}
