% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05_opsum}
\alias{day05_opsum}
\title{add two numbers}
\usage{
day05_opsum(pointer, array, in_buffer, out_buffer)
}
\arguments{
\item{pointer}{pointer}

\item{array}{array}

\item{in_buffer}{input array}

\item{out_buffer}{output array}
}
\description{
add two numbers
}
