% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05_run_computer}
\alias{day05_run_computer}
\title{run diagnostics on INTCODE computer}
\usage{
day05_run_computer(input_buffer, array)
}
\arguments{
\item{input_buffer}{buffer of input values}

\item{array}{INCODE array}
}
\description{
run diagnostics on INTCODE computer
}
