% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05_setv}
\alias{day05_setv}
\title{Set value}
\usage{
day05_setv(value, mode, pointer, base, array)
}
\arguments{
\item{value}{parameter value}

\item{mode}{parameter mode (read from array or show exact value)}

\item{pointer}{pointer to element in array or exact value}

\item{base}{relative base of pointer for parameter mode 2}

\item{array}{array of intcodes}
}
\description{
takes value and puts it into right place in array returning updated array
}
