% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day06.R
\name{day06_build_chains}
\alias{day06_build_chains}
\title{build chains of orbits towards center of everything}
\usage{
day06_build_chains(orbit_map, obj)
}
\arguments{
\item{orbit_map}{object relation map}

\item{obj}{vector of object for which distance needs to be calculated}
}
\description{
Function returns result as a list of object names and
vectors of objects between them and center of everything
}
