% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day06.R
\name{day06_check_splits}
\alias{day06_check_splits}
\title{check that all splits have 2 elements}
\usage{
day06_check_splits(splits)
}
\arguments{
\item{splits}{- list of splits}
}
\description{
check that all splits have 2 elements
}
