% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day06.R
\name{day06_count_orbits}
\alias{day06_count_orbits}
\title{calculate number of orbits for given map and object code}
\usage{
day06_count_orbits(orbit_map, obj)
}
\arguments{
\item{orbit_map}{object relation map}

\item{obj}{vector of object for which distance needs to be calculated}
}
\description{
Function returns result as a list of object names and
number of their direct + indirect orbits
}
