% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day06.R
\name{day06_gather_objects}
\alias{day06_gather_objects}
\title{make list of all objects}
\usage{
day06_gather_objects(orbit_map)
}
\arguments{
\item{orbit_map}{object relation map}
}
\description{
make list of all objects
}
