% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day06.R
\name{day06_rotates_around}
\alias{day06_rotates_around}
\title{Rotates around analyzes object relationships and return list of objects
that point to their direct centers of orbits}
\usage{
day06_rotates_around(orbital_rln)
}
\arguments{
\item{orbital_rln}{of orbital relation codes "A)B" where
A is direct center of rotation and
B rotating object}
}
\description{
Rotates around analyzes object relationships and return list of objects
that point to their direct centers of orbits
}
