% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day07.R
\name{day07_run_curcuit}
\alias{day07_run_curcuit}
\title{Run circuit of amplifiers}
\usage{
day07_run_curcuit(input_instructions, array)
}
\description{
param input_instructions vector of input instructions for every amplifier
}
