% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day07.R
\name{day07_run_feedback_curcuit}
\alias{day07_run_feedback_curcuit}
\title{Run circuit of amplifiers with feedback loop}
\usage{
day07_run_feedback_curcuit(input_instructions, array)
}
\arguments{
\item{input_instructions}{vector of input instructions for every amplifier}

\item{array}{INTCODE array}
}
\description{
Run circuit of amplifiers with feedback loop
}
