% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day08.R
\name{day08_split2chunks}
\alias{day08_split2chunks}
\title{Split vector of values into chunks of fixed size}
\usage{
day08_split2chunks(value, chunk_size)
}
\arguments{
\item{value}{vector of values}

\item{chunk_size}{size of one chunk}
}
\description{
Split vector of values into chunks of fixed size
}
