% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day08.R
\name{day08_top_visible_pixel}
\alias{day08_top_visible_pixel}
\title{determine which pixels are visible for stack of layers}
\usage{
day08_top_visible_pixel(layers)
}
\arguments{
\item{layers}{layers}
}
\description{
determine which pixels are visible for stack of layers
}
