% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day08.R
\name{day08_trace_pixel}
\alias{day08_trace_pixel}
\title{Look at pixel trhough layeers and decide what you see}
\usage{
day08_trace_pixel(prev_val, lower_val)
}
\arguments{
\item{prev_val}{- value of pixel created on previous layers}

\item{lower_val}{- value of pixel in current layer}
}
\description{
Pixel values are 0 - black, 1 - white, 0 - transparent
}
