% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day10.R
\name{day10_contour_around_station}
\alias{day10_contour_around_station}
\title{Identify boundaries to scan with station radar}
\usage{
day10_contour_around_station(station, radius)
}
\arguments{
\item{station}{station coordinates}

\item{radius}{distance (shortest) from station to boundary}
}
\description{
This is square around station
}
