% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day10.R
\name{day10_identify_all}
\alias{day10_identify_all}
\title{Identify all asteroids on a matrix map}
\usage{
day10_identify_all(asteroid_map_text)
}
\arguments{
\item{asteroid_map_text}{text representation of map}
}
\description{
Map is represented by vectod of strings wirh values # and .
}
