% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day10.R
\name{day10_identify_all_polar}
\alias{day10_identify_all_polar}
\title{Identify all polar coordinates of asteroids}
\usage{
day10_identify_all_polar(station, objects)
}
\arguments{
\item{station}{position of station}

\item{objects}{positions of objects represented as list}
}
\description{
Identify all polar coordinates of asteroids
}
