% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day10.R
\name{day10_locate_objects}
\alias{day10_locate_objects}
\title{Locate all asteroids visible to station}
\usage{
day10_locate_objects(station, objects)
}
\arguments{
\item{station}{station coordinates}

\item{objects}{asteroids coordinates}
}
\description{
Locate all asteroids visible to station
}
