% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day10.R
\name{day10_objects_on_contour}
\alias{day10_objects_on_contour}
\title{Identify all objects that are located on contour line}
\usage{
day10_objects_on_contour(objects, contour)
}
\arguments{
\item{objects}{complete set of asterods}

\item{contour}{contour boundaries}
}
\description{
Identify all objects that are located on contour line
}
