% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day10.R
\name{day10_vapor_seq}
\alias{day10_vapor_seq}
\title{asteroid vaporization sequence}
\usage{
day10_vapor_seq(station, asteroids)
}
\arguments{
\item{station}{starion}

\item{asteroids}{asteroids}
}
\description{
asteroid vaporization sequence
}
