% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day10.R
\name{day10_visible}
\alias{day10_visible}
\title{Identify whether asteroid is visible to station}
\usage{
day10_visible(new_object, for_station, among_others)
}
\arguments{
\item{new_object}{asteriod coordinates that needs to be checked}

\item{for_station}{station coordinates}

\item{among_others}{previousy identified asteroids}
}
\description{
Asteroid where station is installed is visible
}
