% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Day_10.R
\name{day10AsteroidViewCount}
\alias{day10AsteroidViewCount}
\title{Asteroid View Count}
\usage{
day10AsteroidViewCount(current_coord, asteroid_list)
}
\arguments{
\item{current_coord}{the coordinate to "look" from}

\item{asteroid_list}{the list of possible asteroids to see}
}
\value{

}
\description{
Asteroid View Count
}
