% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Day_10.R
\name{day10InputParser}
\alias{day10InputParser}
\title{Input Parser for Day 10 of the AOC Challenge}
\usage{
day10InputParser(input_chars)
}
\arguments{
\item{input_chars}{the input character}
}
\value{

}
\description{
Input Parser for Day 10 of the AOC Challenge
}
