% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Day_10.R
\name{day10LengthCalc}
\alias{day10LengthCalc}
\title{Vector Length Calculator}
\usage{
day10LengthCalc(vect)
}
\arguments{
\item{vect}{the vector in question}
}
\value{

}
\description{
Vector Length Calculator
}
