% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Day_10.R
\name{day10MagnitudeCalc}
\alias{day10MagnitudeCalc}
\title{Vector Magnitude Calculator}
\usage{
day10MagnitudeCalc(vect)
}
\arguments{
\item{vect}{the vector in question}
}
\value{

}
\description{
Vector Magnitude Calculator
}
