% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Day_1.R
\name{day1MassToFuelConverter}
\alias{day1MassToFuelConverter}
\title{Mass to fuel conversion function}
\usage{
day1MassToFuelConverter(mass_val)
}
\arguments{
\item{mass_val}{the value of the mass to be converted to fuel}
}
\value{

}
\description{
Mass to fuel conversion function
}
