% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Day_1.R
\name{day1MassToFuelConverter_complete}
\alias{day1MassToFuelConverter_complete}
\title{Recursive Mass to Fuel conversion function}
\usage{
day1MassToFuelConverter_complete(mass_value)
}
\arguments{
\item{mass_value}{the value of the mass to be converted to fuel recursively}
}
\value{

}
\description{
Recursive Mass to Fuel conversion function
}
