% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Day_2_functions.R
\name{day2IntcodeCalculator}
\alias{day2IntcodeCalculator}
\title{Incoder Function}
\usage{
day2IntcodeCalculator(intcode)
}
\arguments{
\item{intcode}{the list of intcode values}
}
\value{
the calculated list of intcode values
}
\description{
Incoder Function
}
