% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Day_2_functions.R
\name{day2StringToIntcode}
\alias{day2StringToIntcode}
\title{String to Intcode converter}
\usage{
day2StringToIntcode(string_intcode)
}
\arguments{
\item{string_intcode}{The inmput string of the intcode value}
}
\value{
an integer vector of the intcode
}
\description{
String to Intcode converter
}
