% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Day_3.R
\name{day3DetermineClosestCrossover}
\alias{day3DetermineClosestCrossover}
\title{Determine the closest crossover}
\usage{
day3DetermineClosestCrossover(input_string1, input_string2,
  vect_reference = c(0, 0))
}
\arguments{
\item{input_string1}{the first input string}

\item{input_string2}{the second input string}

\item{vect_reference}{the point of origin}
}
\description{
Determine the closest crossover
}
