% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Day_3.R
\name{day3DetermineShortestPath}
\alias{day3DetermineShortestPath}
\title{Determine the crossing points of two instruction strings}
\usage{
day3DetermineShortestPath(input_string_1, input_string_2)
}
\arguments{
\item{input_string_1}{the first instruction string}

\item{input_string_2}{the second instruction string}
}
\value{

}
\description{
Determine the crossing points of two instruction strings
}
