% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Day_3.R
\name{day3GetPathFromString}
\alias{day3GetPathFromString}
\title{Generate the path from the string}
\usage{
day3GetPathFromString(input_string)
}
\arguments{
\item{input_string}{the input string}
}
\description{
Generate the path from the string
}
