% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Day_3.R
\name{day3ManhattanDistance}
\alias{day3ManhattanDistance}
\title{Manhattan Distance between two vectors}
\usage{
day3ManhattanDistance(vect_target, vect_reference = c(0, 0))
}
\arguments{
\item{vect_target}{the out of origin vector}

\item{vect_reference}{the origin vector, assumed to be c(0,0)}
}
\description{
Manhattan Distance between two vectors
}
