% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Day_3.R
\name{day3MoveCodeToCoords}
\alias{day3MoveCodeToCoords}
\title{Move codes to Coord}
\usage{
day3MoveCodeToCoords(move_code)
}
\arguments{
\item{move_code}{the move code to transform of the format "[a-zA-Z]\d*"}
}
\value{

}
\description{
Move codes to Coord
}
