% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Day_4.R
\name{day4CheckOneDouble}
\alias{day4CheckOneDouble}
\title{Check for a double which isnt part of a triple or larger}
\usage{
day4CheckOneDouble(vector_of_values)
}
\arguments{
\item{vector_of_values}{the vector to be checked}
}
\value{

}
\description{
Check for a double which isnt part of a triple or larger
}
