% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Day_4.R
\name{day4CheckSequence}
\alias{day4CheckSequence}
\title{Sequence checker}
\usage{
day4CheckSequence(vector_split_sequence, min_val, max_val, part2 = FALSE)
}
\arguments{
\item{vector_split_sequence}{the vector sequence to check}

\item{min_val}{the lower bound}

\item{max_val}{the upper bound}
}
\description{
Sequence checker
}
