% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Day_4.R
\name{day4CheckSequenceDouble}
\alias{day4CheckSequenceDouble}
\title{Check the presence of double numbers}
\usage{
day4CheckSequenceDouble(a_vector_of_values)
}
\arguments{
\item{a_vector_of_values}{The vector to check}
}
\description{
Check the presence of double numbers
}
