% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Day_4.R
\name{day4CheckSequenceLength}
\alias{day4CheckSequenceLength}
\title{Check sequence length}
\usage{
day4CheckSequenceLength(a_vector_of_values, required_length = 6)
}
\arguments{
\item{a_vector_of_values}{the Vector to check}

\item{required_length}{the Length required, defaults to 6}
}
\description{
Check sequence length
}
