% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Day_4.R
\name{day4CheckSequenceMonotonicity}
\alias{day4CheckSequenceMonotonicity}
\title{Monotonicity Check}
\usage{
day4CheckSequenceMonotonicity(a_vector_of_values)
}
\arguments{
\item{a_vector_of_values}{the Vector to check}
}
\description{
Monotonicity Check
}
