% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Day_4.R
\name{day4CheckSequenceRange}
\alias{day4CheckSequenceRange}
\title{Sequence Range checker}
\usage{
day4CheckSequenceRange(a_vector_of_values, min_val, max_val)
}
\arguments{
\item{a_vector_of_values}{the Vector to check}

\item{min_val}{the lower bound}

\item{max_val}{the upper bound}
}
\description{
Sequence Range checker
}
