% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASETS.R
\docType{data}
\name{Day_1_masses}
\alias{Day_1_masses}
\title{Day 1 data}
\format{A list of lists with each Level of the data and each class of column}
\usage{
Day_1_masses
}
\description{
A list of masses for the components of the "spaceship" used to calculate the fuel in
the challenges for day 1
}
\keyword{datasets}
