% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{INTCODE}
\alias{INTCODE}
\title{internal INTCODE computer service data}
\format{An object of class \code{list} of length 2.}
\usage{
INTCODE
}
\description{
internal INTCODE computer service data
}
\keyword{datasets}
