% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DATASET}
\alias{DATASET}
\title{Input DATASETs}
\format{An object of class \code{list} of length 6.}
\usage{
DATASET
}
\description{
This list contains all DATASETs of Advent of Code 2019 event
}
\keyword{datasets}
