% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{day03_part2}
\alias{day03_part2}
\title{Fund find closest intersection of 2 wires defined by instruction sets}
\usage{
day03_part2(instructions1, instructions2)
}
\arguments{
\item{instruction1}{instructions to build path 1}

\item{instruction2}{instructions to build path 2}
}
\description{
Fund find closest intersection of 2 wires defined by instruction sets
}
