% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05_setv}
\alias{day05_setv}
\title{Set value}
\usage{
day05_setv(value, pointer, array)
}
\arguments{
\item{value}{parameter value}

\item{pointer}{pointer to element in array or exact value}

\item{array}{array of intcodes}
}
\description{
Set value
}
