% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day10.R
\name{day10_dec_to_polar}
\alias{day10_dec_to_polar}
\title{Transform coordinates from decart to polar system}
\usage{
day10_dec_to_polar(station, asteroid)
}
\arguments{
\item{station}{station coordinates}

\item{asteroid}{asteroid coordinates}
}
\description{
Transform coordinates from decart to polar system
}
